--- src/civitconfig/data/config_class/defaultconfig.py
@@ -1,7 +1,0 @@
-from ._config import Config
-
-
-class DefaultConfig(Config):
-
-    def __init__(self, *args):
-        super(DefaultConfig, self).__init__(*args)
